+++ src/api.test.js
@@ -4,18 +4,6 @@ describe('api', function() {
   before(function() {
     this.fTwelveAttached = () => this.isAttached(this.fTwelve.el);
   });
-  describe('#setContent()', function() {
-    it('should call setContent and toggle content visibility when a tab is clicked', function() {
-      const tabs = Array.from(this.fTwelve.el.getElementsByClassName('tab'));
-      tabs.forEach((tab) => {
-        const message = 'Clicked tab did not toggle content visibility: ' + tab.innerText;
-        tab.click();
-        assert.strictEqual(this.fTwelve.el.getElementsByClassName('content').length, 1, message);
-        tab.click();
-        assert.strictEqual(this.fTwelve.el.getElementsByClassName('content').length, 0, message);
-      });
-    });
-  });
   describe('#enable()', function() {
     it("should attach to the DOM if 'show' is `true`", function() {
       this.fTwelve.disable();
