+++ src/components/dataTree/Node.test.js
@@ -1,4 +1,4 @@
-import jsx from '../../utilities/jsx';
+import { h } from 'preact';
 import { getNode } from './Tree';
 import Node from './Node';
 import assert from 'assert';
