+++ test/preSetup.js
@@ -11,7 +11,6 @@ register(['.scss']);
  * tool saves copies of the original window.console and window.error at load-time
  */
 global.jsDomCleanup = new JsDomGlobal('', { url: 'http://localhost/' });
-global.SVGElement = () => undefined;
 window.onErrorCallCount = 0;
 window.onerror = () => {
   window.onErrorCallCount++;
